Cambio visual y arreglo bugs

+++ Store/views.py
@@ -1,26 +1,25 @@
 import json
 import requests
-from django.shortcuts import get_object_or_404, render, redirect
-from django.contrib.auth import authenticate, login
-from django.contrib.auth.decorators import login_required
-from django.db import IntegrityError
-from django.contrib.auth import update_session_auth_hash
-from .models import Background, Logo, Usuario
-from .models import Producto, Categoria
-from .models import Carrito
-from django.http import JsonResponse
-from django.http import HttpResponseRedirect
-from django.urls import reverse
-from django.contrib import messages
-from django.shortcuts import render, redirect
-import requests
-import json
-from django.shortcuts import render
 
+from django.contrib import messages
+from django.contrib.auth import authenticate, login, logout, update_session_auth_hash
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.views import PasswordResetView
+from django.contrib.auth import views as auth_views
+from django.contrib.auth import get_user_model
+from django.core.mail import send_mail
+from django.db import IntegrityError
 from django.db.utils import IntegrityError
-from django.contrib.auth import login
-from .models import Detalle
-from django.contrib.auth import logout
+from django.db.models.signals import post_save
+from django.dispatch import receiver
+from django.http import JsonResponse, HttpResponseRedirect
+from django.shortcuts import get_object_or_404, render, redirect
+from django.urls import reverse, reverse_lazy
+from django.shortcuts import render, get_object_or_404
+from .models import Producto, Venta
+from .models import Carrito, Venta, Producto
+from .forms import ProductoSearchForm
+from .models import Background, Logo, Usuario, Producto, Categoria, Carrito, Detalle
 
 
 def vista_categorias(request):
@@ -35,6 +34,8 @@ def vista_categorias(request):
     }
 
     return render(request, 'Store/html/categorias.html', context)
+
+
 
 
 def detalle_categoria(request, id_categoria):
@@ -83,6 +84,9 @@ def cierresesion(request):
 
     return render(request, 'Store/html/cierresesion.html')
 
+
+
+
 def bienvenida(request):
     return render(request, 'Store/html/bienvenida.html')
 
@@ -94,29 +98,39 @@ def editar(request):
 def realizar_pedido(request):
     return render(request, 'Store/html/realizar_pedido.html')
 
+def confirmar_venta(request):
+    if request.method == 'POST':
+        # Asegúrate de que el usuario esté autenticado
+        if not request.user.is_authenticated:
+            return redirect('login')  # Redirigir a la página de inicio de sesión
 
-def loginc(request):
-    site_settings = Logo.objects.first()  
-    background = Background.objects.first() 
+        # Obtener todos los artículos del carrito del usuario
+        carrito = Carrito.objects.filter(usuario=request.user)
 
-    context = {
-        'site_settings': site_settings,
-        'background': background,
-    }
+        # Verificar que el carrito no esté vacío
+        if not carrito.exists():
+            return redirect('carrito_vacio')  # Manejar carrito vacío
 
-    return render(request, 'Store/html/login.html', context)
+        for item in carrito:
+            # Crear la venta por cada artículo en el carrito
+            venta = Venta(
+                usuario=request.user,
+                producto=item.producto,
+                cantidad=item.cantidad,
+                precio_unitario=item.precio_total / item.cantidad,  # Calcula el precio unitario
+                total=item.precio_total
+            )
+            venta.save()  # Guardar la venta en la base de datos
+
+            # Limpiar el artículo del carrito después de la compra
+            item.delete()
+
+        # Redirigir a una página de éxito
+        return redirect('pagina_de_exito')
+
+    return render(request, 'carrito/confirmar.html')
 
 
-def registro(request):
-    site_settings = Logo.objects.first()
-    background = Background.objects.first()
-
-    context = {
-        'site_settings': site_settings,
-        'background': background,
-    }
-
-    return render(request, 'Store/html/registro.html', context)
 
 
 def cambio_contra(request):
@@ -130,6 +144,9 @@ def cambio_contra(request):
 
     return render(request, 'Store/html/cambio_contrasena.html', context)
 
+
+
+
 def compra_finalizada(request):
     return render(request, 'Store/html/compra_finalizada.html')
 
@@ -137,23 +154,80 @@ def lista (request):
     return render(request, 'Store/html/lista_detalles.html')
 
 def index(request):
-    site_settings = Logo.objects.first() 
-    background = Background.objects.first()  
+    site_settings = Logo.objects.first()
+    background = Background.objects.first()
+
+    # Filtrar las categorías que deben aparecer en el carrusel
+    categorias = Categoria.objects.filter(visible_en_carrusel=True)
+
+    # Obtener productos más vendidos
+    productos_mas_vendidos = Venta.obtener_productos_mas_vendidos()
+
+    # Obtener productos más nuevos
+    productos_mas_nuevos = Producto.objects.order_by('-idProducto')[:5]
 
     context = {
-        'site_settings': site_settings,  
-        'background': background,  
+        'site_settings': site_settings,
+        'background': background,
+        'categorias': categorias,
+        'productos_mas_vendidos': productos_mas_vendidos,  # Añadir productos más vendidos al contexto
+        'productos_mas_nuevos': productos_mas_nuevos,  
     }
 
     return render(request, 'Store/index.html', context)
 
 
 
+def producto_detail(request, idProducto):
+    producto = get_object_or_404(Producto, idProducto=idProducto)
+    context = {
+        'producto': producto
+    }
+    return render(request, 'Store/producto_detail.html', context)
+
+def registro(request):
+    site_settings = Logo.objects.first()
+    background = Background.objects.first()
+
+    context = {
+        'site_settings': site_settings,
+        'background': background,
+    }
+
+    return render(request, 'Store/html/registro.html', context)
+
+
+class CustomPasswordResetView(PasswordResetView):
+    email_template_name = 'registration/password_reset_email.html'
+    subject_template_name = 'registration/password_reset_subject.txt'
+    template_name = 'registration/password_reset_form.html'
+    success_url = reverse_lazy('password_reset_done')
+    from_email = 'francopavez1@outlook.com'
+
+    def get_users(self, email):
+        UserModel = get_user_model()
+        # Busca usuarios activos con ese correo
+        active_users = UserModel.objects.filter(email=email, is_active=True)
+        if not active_users.exists():
+            print(f"No se encontraron usuarios activos con el correo: {email}")
+        else:
+            print(f"Usuarios encontrados: {active_users}")
+        return (u for u in active_users)
+
+class CustomPasswordResetCompleteView(auth_views.PasswordResetCompleteView):
+    template_name = 'registration/password_reset_complete.html'
+    success_url = reverse_lazy('iniciar_sesion') 
+
+
 #CARRITO
 def carrito(request):
-  
-    carrito_items = Carrito.objects.filter(usuario=request.user)
+    site_settings = Logo.objects.first()
+    background = Background.objects.first()
+    
+    carrito_items = Carrito.objects.filter()
     context = {
+        'site_settings': site_settings,
+        'background': background,
         'carrito_items': carrito_items,
     }
     return render(request, 'Store/html/carrito.html', context)
@@ -288,6 +362,10 @@ def registrar_usuario(request):
 
 # INICAR SESION
 def iniciar_sesion(request):
+    # Obtén los objetos 'Logo' y 'Background' antes de procesar la solicitud
+    site_settings = Logo.objects.first()
+    background = Background.objects.first()
+
     if request.method == 'POST':
         username = request.POST['user']
         password = request.POST['pass']
@@ -295,7 +373,7 @@ def iniciar_sesion(request):
 
         # Verificar el Captcha
         cap_url = "https://www.google.com/recaptcha/api/siteverify"
-        cap_secret = "6LfxrGooAAAAAAZX1lgSKeFjKZ1s4INgXReeBuXJ" 
+        cap_secret = "6LfxrGooAAAAAAZX1lgSKeFjKZ1s4INgXReeBuXJ"
         cap_data = {"secret": cap_secret, "response": captcha_token}
         cap_server_response = requests.post(url=cap_url, data=cap_data)
         cap_json = json.loads(cap_server_response.text)
@@ -307,40 +385,73 @@ def iniciar_sesion(request):
                 return redirect('index')
             else:
                 error_message = "Nombre de usuario o contraseña incorrectos."
-                return render(request, 'Store/html/login.html', {'error_message': error_message})
+                return render(request, 'Store/html/login.html', {
+                    'error_message': error_message,
+                    'site_settings': site_settings,
+                    'background': background
+                })
         else:
             error_message = "Por favor, completa el Captcha correctamente."
-            return render(request, 'Store/html/login.html', {'error_message': error_message})
+            return render(request, 'Store/html/login.html', {
+                'error_message': error_message,
+                'site_settings': site_settings,
+                'background': background
+            })
     
-    return render(request, 'Store/html/login.html')
+    return render(request, 'Store/html/login.html', {
+        'site_settings': site_settings,
+        'background': background
+    })
 
+
+
+def buscar_producto(request):
+    form = ProductoSearchForm()
+    productos = None
+    site_settings = Logo.objects.first()
+    background = Background.objects.first()
+
+    if 'query' in request.GET:
+        query = request.GET.get('query')
+        productos = Producto.objects.filter(
+            nombreProducto__icontains=query
+        ) | Producto.objects.filter(
+            descripcionProducto__icontains=query
+        ) | Producto.objects.filter(
+            precioProducto__icontains=query
+        )
+     
+    return render(request, 'Store/html/search_results.html', {
+        'form': form,
+        'site_settings': site_settings,
+        'background': background,
+        'productos': productos
+    })
 
 
 #AGREGAR UNIDADES AL CARRITO
 
-def agregar_al_carrito(request, producto_id):
-    producto = Producto.objects.get(pk=producto_id)
+def agregar_al_carrito(request, idProducto):
+    producto = get_object_or_404(Producto, idProducto=idProducto)
+    
+    # Obtener la cantidad del formulario
+    cantidad = int(request.POST.get('cantidad', 1))
 
-    if request.user.is_authenticated:
-        cantidad = int(request.POST.get('cantidad', 1))
+    # Lógica para agregar al carrito
+    carrito_item, created = Carrito.objects.get_or_create(
+        usuario=request.user,
+        producto=producto,
+        defaults={'cantidad': 0}
+    )
+    carrito_item.cantidad += cantidad
+    carrito_item.precio_total = carrito_item.cantidad * producto.precioProducto
+    carrito_item.save()
 
-        if cantidad <= producto.stockProd:
-            precio_total = producto.precioProducto * cantidad
-            carrito, created = Carrito.objects.get_or_create(usuario=request.user, producto=producto)
+    # Verificar si el artículo se agrega correctamente
+    print(f'Producto agregado: {producto.nombreProducto}, Cantidad: {carrito_item.cantidad}')
 
-            if (carrito.cantidad + cantidad) <= producto.stockProd:
-                carrito.cantidad += cantidad
-                carrito.precio_total = precio_total
-                carrito.save()
-                messages.success(request, "Producto agregado al carrito correctamente.")
-            else:
-                messages.error(request, "La cantidad solicitada supera el stock disponible en el carrito.")
-        else:
-            messages.error(request, "La cantidad solicitada supera el stock disponible.")
-    else:
-        messages.warning(request, "Debes iniciar sesión para agregar productos al carrito.")
-
-    return HttpResponseRedirect(request.META.get('HTTP_REFERER')) # se regresa a la misma pagian donde esta
+    # Redirigir a la página anterior
+    return redirect(request.META.get('HTTP_REFERER', 'index'))
 
 
 
@@ -362,18 +473,31 @@ def vaciar_carrito(request):
 
 
 
-#FUNCION PARA FINALIZAR COMPRA
 def finalizar_compra(request):
-    carrito_items = Carrito.objects.filter(usuario=request.user)
+    if request.method == 'POST':
+        if request.user.is_authenticated:
+            # Procesa la dirección del envío
+            direccion = request.POST.get('direccion')
+            carrito_items = Carrito.objects.filter(usuario=request.user)
 
-    for item in carrito_items:
-        producto = item.producto
-        producto.stockProd -= item.cantidad
-        producto.save()
+            if carrito_items.exists():
+                for item in carrito_items:
+                    # Crea una nueva venta por cada artículo en el carrito
+                    venta = Venta(
+                        usuario=request.user,
+                        producto=item.producto,
+                        cantidad=item.cantidad,
+                        precio_unitario=item.producto.precioProducto,
+                        total=item.precio_total,
+                        imagen_producto=item.producto.imagen  # Guardar la imagen del producto
+                    )
+                    venta.save()
 
-    
-    carrito_items.delete()
+                # Vaciar el carrito después de la compra
+                carrito_items.delete()
 
-    return redirect('compra_finalizada')  
+                return redirect('index') 
+            else:
+                return redirect('carrito_vacio')  
 
-
+    return render(request, 'carrito/finalizar_compra.html')
